v1.0.6

+++ src/features/Verification/validators.ts
@@ -5,7 +5,7 @@ import { HandleError } from '../../responses/error/HandleError'
 import crypto from 'crypto'
 
 export const verificationValidate = async (req: Request, res: Response, next: NextFunction) => {
-    const path:string = req.params.path
+    const path: string = req.params.path
 
     const verification = await Verification.findOne({ path })
     if (!verification) return next(HandleError.NotFound('Path expired'))
@@ -19,13 +19,15 @@ export const verificationValidate = async (req: Request, res: Response, next: Ne
 }
 
 export const verificationConfirm = async (req: Request, res: Response, next: NextFunction) => {
-    const code:string = req.body.code
-    const path:string = req.params.path
+    const code: string = req.body.code
+    const path: string = req.params.path
+
+    if (code.length > 8) return next(HandleError.NotAcceptable('Code is 8 characters long'))
 
     const verification = await Verification.findOneAndUpdate({ path }, { $inc: { tries: -1 } }, { new: true })
     if (!verification) return next(HandleError.NotFound('Path expired'))
 
-    const yourCodeHashed:string = crypto.createHash('sha256').update(code).digest('hex')
+    const yourCodeHashed: string = crypto.createHash('sha256').update(code).digest('hex')
 
     if (verification.codeHashed !== yourCodeHashed && verification.tries <= 0) {
         await Verification.findOneAndRemove({ path })
